+++ backend/src/main/java/com/grepp/smartwatcha/app/controller/api/recommend/payload/MovieRecommendLatestResponse.java
@@ -1,5 +1,4 @@
-package com.grepp.smartwatcha.infra.response.recommend;
-
+package com.grepp.smartwatcha.app.controller.api.recommend.payload;
 
 import com.grepp.smartwatcha.infra.jpa.entity.MovieEntity;
 import lombok.AllArgsConstructor;
@@ -10,7 +9,7 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class MovieRecommendHighestRatedResponse {
+public class MovieRecommendLatestResponse {
 
     private Long id;
     private String title;
@@ -18,10 +17,10 @@ public class MovieRecommendHighestRatedResponse {
     private String country;
     private String poster;
     private Double avgScore;
-    private List<String> genres;
+    private List<String> genre;
 
-    public static MovieRecommendHighestRatedResponse from(MovieEntity movie, Double avgScore, List<String> genres) {
-        return new MovieRecommendHighestRatedResponse(
+    public static MovieRecommendLatestResponse from(MovieEntity movie, Double avgScore, List<String> genres) {
+        return new MovieRecommendLatestResponse(
                 movie.getId(),
                 movie.getTitle(),
                 movie.getReleaseDate(),
