+++ backend/src/main/java/com/grepp/smartwatcha/app/model/recommend/RecommendLatestMovieService.java
@@ -1,7 +1,7 @@
-package com.grepp.smartwatcha.app.service;
+package com.grepp.smartwatcha.app.model.recommend;
 
-import com.grepp.smartwatcha.app.service.recommend.RecommendLatestRatedJpaService;
-import com.grepp.smartwatcha.infra.response.recommend.MovieRecommendLatestResponse;
+import com.grepp.smartwatcha.app.model.recommend.service.RecommendLatestRatedJpaService;
+import com.grepp.smartwatcha.app.controller.api.recommend.payload.MovieRecommendLatestResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
