+++ backend/src/main/java/com/grepp/smartwatcha/app/model/recommend/repository/MovieGenreNeo4jRepository.java
@@ -1,4 +1,4 @@
-package com.grepp.smartwatcha.infra.neo4j.recommend;
+package com.grepp.smartwatcha.app.model.recommend.repository;
 
 import com.grepp.smartwatcha.infra.neo4j.node.MovieNode;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
