+++ backend/src/main/java/com/grepp/smartwatcha/app/model/recommend/repository/RecommendHighestRatedMovieJpaRepository.java
@@ -1,4 +1,4 @@
-package com.grepp.smartwatcha.infra.jpa.recommend;
+package com.grepp.smartwatcha.app.model.recommend.repository;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
