+++ backend/src/main/java/com/grepp/smartwatcha/app/model/recommend/service/RecommendLatestRatedNeo4jService.java
@@ -1,8 +1,8 @@
-package com.grepp.smartwatcha.app.service.recommend;
+package com.grepp.smartwatcha.app.model.recommend.service;
 
 import com.grepp.smartwatcha.infra.neo4j.node.GenreNode;
 import com.grepp.smartwatcha.infra.neo4j.node.MovieNode;
-import com.grepp.smartwatcha.infra.neo4j.recommend.MovieGenreNeo4jRepository;
+import com.grepp.smartwatcha.app.model.recommend.repository.MovieGenreNeo4jRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
